More Metallurgy 3 compatibility
I tried to get metallurgic infuser to work with the new alloys from
Metallurgy, but it was much harder than anticipated and involved major
rewrite on how infusion types work.

InfusionType changed from enum to class, allowing new types to be added
runtime. List of InfuseTypes is maintained at Metallurgy.java, as a map
with name string as key and type as value.

The old enum types are now added in Metallurgy.java, exept for Bio,
which is from Generators.

GUI defaults to the diamond bar if a specific bar for the infuse type
doesn't exist.

The NEI recipe viewer also works, and displays all the new recipes..

Also added enrichment chamber reward recipes for Metallurgy utility ores
that drop items other than the ore itself, akin to redstone.