Fix for digital miner tile entity going out of control on full inventory
When the digital miner is active and mining, if there is no room for the
drops of the currently designated block mined, it can get caught in an
infinite while loop condition which will freeze the game. This is a fix
for this behavior, as well as a bit of optimization and change in
behavior of the digital miner.

Currently the digital miner, when it encounters this condition, it will
keep drawing power and will try to move on to mining the next block in
the list. With a full inventory condition, this can be suboptimal, as
potentially it could be going through the entire block list each time
looking for a block that it can mine, and this will increase the
processing time of the tile entity. This commit changes this behavior,
now it will get paused when the last inventory slot gets filled, and
will cease drawing power. So long as there is somewhere for the drops to
go, the miner will keep mining as expected.

Also there is a bit of a fix to the while loop where if the while loop
has reached the end of the space it should search, it will stop running.