Fixed some Electric Chest bugs
Electric Chest was not updating when it was (un)locked, meaning that
mechanical pipes would not (dis)connect and reconnect.

Also noticed that Electric Chest won't accept transporters connecting to
the bottom, which I thought was odd, however it looked intentional so I
left it alone. (..?)

Also made it so electric chests will not open if there is a solid block
above them.