Fixes issue #1589, but note the complication.
Fixes issue: Fluidic Plenisher doesn't detect block update directly
below it #1589

Fluidic Plenisher will now reset its' calculations after a neighboring
block updates - in the same way it does when you shift + right click
with a configurator.  In the case of the issue, it will update when you
destroy/remove the obsidian.

However, this can cause problems. Say you have a lake that you have
filled using the Plenisher. Should any block next to the Plenisher
update it will cause a reset, and the Plenisher will replace all the
blocks in the lake with new ones. You essentially re-fill an already
full lake. Obviously this is very wasteful and not particularly wanted.

A possible fix to this could be that the Plenisher does not place the
block on source blocks of the same type? This would also mean "topping
up" a pre-made lake will not waste lava.

Thought I would see what you thought, and then maybe I will also
implement the change above if you think it is a good idea?