Down below 2500 errors now, with the following packages excluded: * mekanism.client.nei * mekanism.client.render.block * mekanism.common.multipart
multipart and NEI are on ChickenBones, render.block is probably not going to be a thing any more, but I'm keeping it around for now for reference.

Our api is clear and I've done most of the basic find-replace work now, I think.
Coord4D has been converted to a subclass of BlockPos now that the latter is prevalent, for convenience.