Sorting buttons now work
Got the sorting buttons to work. Clicking on the arrow buttons on a
filter panel will now move that filter up or down in the filter list.
Without having made any changed to how filters are processed, filters
higher in the list will naturally be processed before later filters,
thereby making it like a priority system.
For example, if an item matches two filters, the higher priority filter
will take that item before the lower filter, in that way seperating a
specific item from a broader (wildcard) specified filter.