Only allow same fluids in active fluidtank towers.
Implements feature #4165
If the current fluidtank is active and contains no fluid, check if the fluidtank underneath has the same fluid.
If it's the same as the one being inserted, accept the fluid. If they aren't the same, do not accept the fluid.
If the fluidtank underneath will execute the same checks if active.

Original behaviour is the same with non active fluidtanks.