acceptEnergy/pullEnergy: return 0 if machine is full or empty
Prior to this change, the machines always updated their energy, which
involved a side check and a chunk markDirty; even if the machine is
receiving or giving 0 energy.
Right upfront, check if the machine needs no energy or has none to give,
and if so return 0.
I observed greatly increased server tick performance after this change.