Fix personal chests
Current behaviour is as follows:
The top 6 rows (the real slots for personal chest) disallow placing personal chests in them through SlotPersonalChest.
The other rows (the real player inventory) do allow placing/moving personal chests.
One exception to this is when the personal chest is opened from the hotbar, it disallowes moving the opened chest in the hotbar.