Improvements to fix for #15; smooth out sounds and detect ongoing attempts to muffle
The previous commit made the assumption that detection of sound being complete was fast enough
that restarting a well-designed sound would be seamless -- this is not true. As such, had
to change sounds to run for as long as machine is active on repeat with ongoing detection of
muffling attempts. In addition, if sound was INITIALLY muffled, it would never have been unmuffled,
so this also fixes that problem.