Overhaul how logistics networks estimate what can fit in the network (#124)
This completely changes the prediction logic when items are inserted into a network. In our testing it works well for every mod, with the occasional issue from Storage Drawers. Their interface for IItemHandler is a little odd and causes problems when they get very full and the network is slow.

* Completely overhaul how prediction of paths is done; use IItemHandler's exclusively
* Add exponential backoff when pulling items, up to 40 ticks
* Cache getSlotLimit; add TODO for investigating why we have a for loop
* Do proper simulation insert when dest slot is empty
* Add better handling of new stacks and break loop once stack is fully used up